+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.4.1] - WIP
+## [1.4.1] - 2026-02-08
 
 ### Added
 - New Shell Companion menu item (Shift+Cmd+S) to quickly create a shell companion for the active agent